validation and policies

+++ app/Http/Controllers/AcademicDevelopment/CollegeDepartmentAwardController.php
@@ -19,6 +19,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', CollegeDepartmentAward::class);
+
         $college_department_awards = CollegeDepartmentAward::where('user_id', auth()->id())->get();
         return view('academic-development.college-department-award.index', compact('college_department_awards'));
     }
@@ -30,6 +32,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', CollegeDepartmentAward::class);
+
         $awardFields = DB::select("CALL get_academic_development_fields_by_form_id(6)");
 
         return view('academic-development.college-department-award.create', compact('awardFields'));
@@ -43,6 +47,17 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', CollegeDepartmentAward::class);
+
+        $request->validate([
+            'name_of_award' => 'required',
+            'certifying_body' => 'required',
+            // 'place' => '',
+            'date' => 'required|date',
+            'level' => 'required',
+            // 'description' => 'required',
+        ]);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $college_department_award = CollegeDepartmentAward::create($input);
@@ -82,6 +97,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function show(CollegeDepartmentAward $college_department_award)
     {
+        $this->authorize('view', CollegeDepartmentAward::class);
+
         $awardFields = DB::select("CALL get_academic_development_fields_by_form_id(6)");
 
         $documents = CollegeDepartmentAwardDocument::where('college_department_award_id', $college_department_award->id)->get()->toArray();
@@ -99,6 +116,8 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function edit(CollegeDepartmentAward $college_department_award)
     {
+        $this->authorize('update', CollegeDepartmentAward::class);
+
         $awardFields = DB::select("CALL get_academic_development_fields_by_form_id(6)");
 
         $documents = CollegeDepartmentAwardDocument::where('college_department_award_id', $college_department_award->id)->get()->toArray();
@@ -117,6 +136,17 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function update(Request $request, CollegeDepartmentAward $college_department_award)
     {
+        $this->authorize('update', CollegeDepartmentAward::class);
+
+        $request->validate([
+            'name_of_award' => 'required',
+            'certifying_body' => 'required',
+            // 'place' => '',
+            'date' => 'required|date',
+            'level' => 'required',
+            // 'description' => 'required',
+        ]);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $college_department_award->update($input);
@@ -155,12 +185,16 @@ class CollegeDepartmentAwardController extends Controller
      */
     public function destroy(CollegeDepartmentAward $college_department_award)
     {
+        $this->authorize('delete', CollegeDepartmentAward::class);
+
         CollegeDepartmentAwardDocument::where('college_department_award_id', $college_department_award->id)->delete();
         $college_department_award->delete();
         return redirect()->route('college-department-award.index')->with('award_success', 'Your Accomplishment in Awards and Recognition Received by the College and Department has been deleted.');
     }
 
     public function removeDoc($filename){
+        $this->authorize('delete', CollegeDepartmentAward::class);
+
         CollegeDepartmentAwardDocument::where('filename', $filename)->delete();
         return true;
     }
