+++ app/Http/Controllers/AcademicDevelopment/ReferenceController.php
@@ -21,6 +21,8 @@ class ReferenceController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Reference::class);
+
         $allRtmmi = Reference::where('user_id', auth()->id())
                                         ->join('dropdown_options', 'dropdown_options.id', 'references.category')
                                         ->select('references.*', 'dropdown_options.name as category_name')
@@ -36,6 +38,8 @@ class ReferenceController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Reference::class);
+
         $referenceFields = DB::select("CALL get_academic_development_fields_by_form_id(1)");
 
         $colleges = College::all();
@@ -50,6 +54,8 @@ class ReferenceController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Reference::class);
+
         $request->validate([
             'category' => 'required',
             'level' => 'required',
@@ -65,7 +71,7 @@ class ReferenceController extends Controller
             // 'copyright_regi_no' => ''
             'college_id' => 'required',
             'department_id' => 'required',
-            'description' => 'required',
+            // 'description' => 'required',
         ]);
 
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
@@ -112,6 +118,8 @@ class ReferenceController extends Controller
      */
     public function show(Reference $rtmmi)
     {
+        $this->authorize('view', Reference::class);
+
         $referenceDocuments = ReferenceDocument::where('reference_id', $rtmmi->id)->get()->toArray();
 
         $category = DB::select("CALL get_dropdown_name_by_id(".$rtmmi->category.")");
@@ -133,6 +141,8 @@ class ReferenceController extends Controller
      */
     public function edit(Reference $rtmmi)
     {
+        $this->authorize('update', Reference::class);
+
         $referenceFields = DB::select("CALL get_academic_development_fields_by_form_id(1)");
 
         $referenceDocuments = ReferenceDocument::where('reference_id', $rtmmi->id)->get()->toArray();
@@ -160,6 +170,8 @@ class ReferenceController extends Controller
      */
     public function update(Request $request, Reference $rtmmi)
     {
+        $this->authorize('update', Reference::class);
+
         $request->validate([
             'category' => 'required',
             'level' => 'required',
@@ -175,7 +187,7 @@ class ReferenceController extends Controller
             // 'copyright_regi_no' => ''
             'college_id' => 'required',
             'department_id' => 'required',
-            'description' => 'required',
+            // 'description' => 'required',
         ]);
 
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
@@ -222,6 +234,8 @@ class ReferenceController extends Controller
      */
     public function destroy(Reference $rtmmi)
     {
+        $this->authorize('delete', Reference::class);
+
         $rtmmi->delete();
         ReferenceDocument::where('reference_id', $rtmmi->id)->delete();
 
@@ -235,6 +249,9 @@ class ReferenceController extends Controller
     }
 
     public function removeDoc($filename){
+        
+        $this->authorize('delete', Reference::class);
+
         ReferenceDocument::where('filename', $filename)->delete();
         // Storage::delete('documents/'.$filename);
         return true;
