+++ app/Http/Controllers/AcademicDevelopment/StudentAwardController.php
@@ -19,6 +19,8 @@ class StudentAwardController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', StudentAward::class);
+
         $student_awards = StudentAward::where('user_id', auth()->id())->get();
         return view('academic-development.student-awards.index', compact('student_awards'));
     }
@@ -30,6 +32,8 @@ class StudentAwardController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', StudentAward::class);
+
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(3)");
 
         return view('academic-development.student-awards.create', compact('studentFields'));
@@ -43,6 +47,17 @@ class StudentAwardController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', StudentAward::class);
+
+        $request->validate([
+            'name_of_award' => 'required',
+            'certifying_body' => 'required',
+            // 'place' => '',
+            'date' => 'required|date',
+            'level' => 'required',
+            // 'description' => 'required',
+        ]);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $student_award = StudentAward::create($input);
@@ -82,6 +97,8 @@ class StudentAwardController extends Controller
      */
     public function show(StudentAward $student_award)
     {
+        $this->authorize('view', StudentAward::class);
+
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(3)");
 
         $documents = StudentAwardDocument::where('student_award_id', $student_award->id)->get()->toArray();
@@ -99,6 +116,8 @@ class StudentAwardController extends Controller
      */
     public function edit(StudentAward $student_award)
     {
+        $this->authorize('update', StudentAward::class);
+
         $studentFields = DB::select("CALL get_academic_development_fields_by_form_id(3)");
 
         $documents = StudentAwardDocument::where('student_award_id', $student_award->id)->get()->toArray();
@@ -117,6 +136,17 @@ class StudentAwardController extends Controller
      */
     public function update(Request $request, StudentAward $student_award)
     {
+        $this->authorize('update', StudentAward::class);
+        
+        $request->validate([
+            'name_of_award' => 'required',
+            'certifying_body' => 'required',
+            // 'place' => '',
+            'date' => 'required|date',
+            'level' => 'required',
+            // 'description' => 'required',
+        ]);
+        
         $input = $request->except(['_token', '_method', 'document']);
 
         $student_award->update($input);
@@ -155,12 +185,16 @@ class StudentAwardController extends Controller
      */
     public function destroy(StudentAward $student_award)
     {
+        $this->authorize('delete', StudentAward::class);
+
         StudentAwardDocument::where('student_award_id', $student_award->id)->delete();
         $student_award->delete();
         return redirect()->route('student-award.index')->with('student_success', 'Your accomplishment in Student Award and Recognition has been deleted.');
     }
 
     public function removeDoc($filename){
+        $this->authorize('delete', StudentAward::class);
+
         StudentAwardDocument::where('filename', $filename)->delete();
         return true;
     }
