+++ app/Http/Controllers/AcademicDevelopment/TechnicalExtensionController.php
@@ -19,6 +19,8 @@ class TechnicalExtensionController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', TechnicalExtension::class);
+
         $technical_extensions = TechnicalExtension::where('user_id', auth()->id())->get();
         return view('academic-development.technical-extension.index', compact('technical_extensions'));
     }
@@ -30,6 +32,8 @@ class TechnicalExtensionController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', TechnicalExtension::class);
+
         $extensionFields = DB::select("CALL get_academic_development_fields_by_form_id(7)");
 
         return view('academic-development.technical-extension.create', compact('extensionFields'));
@@ -43,6 +47,23 @@ class TechnicalExtensionController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', TechnicalExtension::class);
+
+        $request->validate([
+            'moa_code' => 'required',
+            'program_title' => 'required',
+            // 'project_title' => '',
+            // 'activity_title' => '',
+            // 'name_of_adoptor' => '',
+            'classification_of_adoptor' => 'required',
+            // 'nature_of_business_enterprise' => '',
+            'has_businesses' => 'required',
+            'is_borrowed' => 'required',
+            'currency_total_profit' => 'required',
+            'total_profit' => 'numeric',
+            // 'description' => 'required',
+        ]);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $technical_extension = TechnicalExtension::create($input);
@@ -82,6 +103,8 @@ class TechnicalExtensionController extends Controller
      */
     public function show(TechnicalExtension $technical_extension)
     {
+        $this->authorize('view', TechnicalExtension::class);
+
         $extensionFields = DB::select("CALL get_academic_development_fields_by_form_id(7)");
 
         $documents = TechnicalExtensionDocument::where('technical_extension_id', $technical_extension->id)->get()->toArray();
@@ -99,6 +122,8 @@ class TechnicalExtensionController extends Controller
      */
     public function edit(TechnicalExtension $technical_extension)
     {
+        $this->authorize('update', TechnicalExtension::class);
+
         $extensionFields = DB::select("CALL get_academic_development_fields_by_form_id(7)");
 
         $documents = TechnicalExtensionDocument::where('technical_extension_id', $technical_extension->id)->get()->toArray();
@@ -117,6 +142,23 @@ class TechnicalExtensionController extends Controller
      */
     public function update(Request $request, TechnicalExtension $technical_extension)
     {
+        $this->authorize('update', TechnicalExtension::class);
+
+        $request->validate([
+            'moa_code' => 'required',
+            'program_title' => 'required',
+            // 'project_title' => '',
+            // 'activity_title' => '',
+            // 'name_of_adoptor' => '',
+            'classification_of_adoptor' => 'required',
+            // 'nature_of_business_enterprise' => '',
+            'has_businesses' => 'required',
+            'is_borrowed' => 'required',
+            'currency_total_profit' => 'required',
+            'total_profit' => 'numeric',
+            // 'description' => 'required',
+        ]);
+
         $input = $request->except(['_token', '_method', 'document']);
 
         $technical_extension->update($input);
@@ -155,12 +197,16 @@ class TechnicalExtensionController extends Controller
      */
     public function destroy(TechnicalExtension $technical_extension)
     {
+        $this->authorize('delete', TechnicalExtension::class);
+
         TechnicalExtensionDocument::where('technical_extension_id', $technical_extension->id)->delete();
         $technical_extension->delete();
         return redirect()->route('technical-extension.index')->with('award_success', 'Your Accomplishment in Technical Extension Programs/ Projects/ Activities has been deleted.');
     }
 
     public function removeDoc($filename){
+        $this->authorize('delete', TechnicalExtension::class);
+
         TechnicalExtensionDocument::where('filename', $filename)->delete();
         return true;
     }
