+++ app/Http/Controllers/AcademicDevelopment/ViableProjectController.php
@@ -19,6 +19,8 @@ class ViableProjectController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', ViableProject::class);
+
         $viable_projects = ViableProject::where('user_id', auth()->id())->get();
         return view('academic-development.viable-project.index', compact('viable_projects'));
     }
@@ -30,6 +32,8 @@ class ViableProjectController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', ViableProject::class);
+
         $projectFields = DB::select("CALL get_academic_development_fields_by_form_id(5)");
 
         return view('academic-development.viable-project.create', compact('projectFields'));
@@ -43,11 +47,28 @@ class ViableProjectController extends Controller
      */
     public function store(Request $request)
     {
-        $input = $request->except(['_token', '_method', 'document']);
+        $this->authorize('create', ViableProject::class);
+
+        $request->validate([
+            'name' => 'required',
+            'currency_revenue' => 'required',
+            'revenue' => 'numeric',
+            'currency_cost' => 'required',
+            'cost' => 'numeric',
+            'start_date' => 'required|date',
+            'rate_of_return' => 'required|numeric',
+            // 'description' => 'required',
+        ]);
+
+        $input = $request->except(['_token', '_method', 'document', 'rate_of_return']);
 
         $viable_project = ViableProject::create($input);
         $viable_project->update(['user_id' => auth()->id()]);
         
+        $return_rate = ($request->input('rate_of_return') / 100 );
+        
+        $viable_project->update(['rate_of_return' => $return_rate]);
+
         if($request->has('document')){
             
             $documents = $request->input('document');
@@ -82,6 +103,8 @@ class ViableProjectController extends Controller
      */
     public function show(ViableProject $viable_project)
     {
+        $this->authorize('view', ViableProject::class);
+
         $projectFields = DB::select("CALL get_academic_development_fields_by_form_id(5)");
 
         $documents = ViableProjectDocument::where('viable_project_id', $viable_project->id)->get()->toArray();
@@ -99,6 +122,8 @@ class ViableProjectController extends Controller
      */
     public function edit(ViableProject $viable_project)
     {
+        $this->authorize('update', ViableProject::class);
+
         $projectFields = DB::select("CALL get_academic_development_fields_by_form_id(5)");
 
         $documents = ViableProjectDocument::where('viable_project_id', $viable_project->id)->get()->toArray();
@@ -117,6 +142,19 @@ class ViableProjectController extends Controller
      */
     public function update(Request $request, ViableProject $viable_project)
     {
+        $this->authorize('update', ViableProject::class);
+
+        $request->validate([
+            'name' => 'required',
+            'currency_revenue' => 'required',
+            'revenue' => 'numeric',
+            'currency_cost' => 'required',
+            'cost' => 'numeric',
+            'start_date' => 'required|date',
+            'rate_of_return' => 'required|numeric',
+            // 'description' => 'required',
+        ]);
+        
         $input = $request->except(['_token', '_method', 'document']);
 
         $viable_project->update($input);
@@ -155,12 +193,16 @@ class ViableProjectController extends Controller
      */
     public function destroy(ViableProject $viable_project)
     {
+        $this->authorize('delete', ViableProject::class);
+
         ViableProjectDocument::where('viable_project_id', $viable_project->id)->delete();
         $viable_project->delete();
         return redirect()->route('viable-project.index')->with('project_success', 'Your accomplishment in Viable Demonstration Project has been deleted.');
     }
 
     public function removeDoc($filename){
+        $this->authorize('delete', ViableProject::class);
+
         ViableProjectDocument::where('filename', $filename)->delete();
         return true;
     }
