+++ app/Http/Controllers/ExtensionPrograms/MobilityController.php
@@ -21,6 +21,8 @@ class MobilityController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Mobility::class);
+
         $mobilities = Mobility::where('user_id', auth()->id())->orderBy('updated_at', 'desc')->get();
         return view('extension-programs.mobility.index', compact('mobilities'));
     }
@@ -32,6 +34,8 @@ class MobilityController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Mobility::class);
+
         $mobilityFields = DB::select("CALL get_extension_program_fields_by_form_id('6')");
 
         $colleges = College::all();
@@ -46,6 +50,21 @@ class MobilityController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Mobility::class);
+
+        $request->validate([
+            'nature_of_engagement' => 'required',
+            'type' => 'required',
+            'host_name' => 'required',
+            // 'host_address' => '',
+            'mobility_description' => 'required',
+            'start_date' => 'required|date',
+            'end_date' => 'required|date|after_or_equal:start_date',
+            'college_id' => 'required',
+            'department_id' => 'required',
+            // 'description' => 'required',
+        ]);
+
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $mobility = Mobility::create($input);
@@ -86,6 +105,8 @@ class MobilityController extends Controller
      */
     public function show(Mobility $mobility)
     {
+        $this->authorize('view', Mobility::class);
+
         $mobilityFields = DB::select("CALL get_extension_program_fields_by_form_id('6')");
 
         $documents = MobilityDocument::where('mobility_id', $mobility->id)->get()->toArray();
@@ -103,6 +124,8 @@ class MobilityController extends Controller
      */
     public function edit(Mobility $mobility)
     {
+        $this->authorize('update', Mobility::class);
+
         $mobilityFields = DB::select("CALL get_extension_program_fields_by_form_id('6')");
 
         $collegeAndDepartment = Department::join('colleges', 'colleges.id', 'departments.college_id')
@@ -128,6 +151,21 @@ class MobilityController extends Controller
      */
     public function update(Request $request, Mobility $mobility)
     {
+        $this->authorize('update', Mobility::class);
+
+        $request->validate([
+            'nature_of_engagement' => 'required',
+            'type' => 'required',
+            'host_name' => 'required',
+            // 'host_address' => '',
+            'mobility_description' => 'required',
+            'start_date' => 'required|date',
+            'end_date' => 'required|date|after_or_equal:start_date',
+            'college_id' => 'required',
+            'department_id' => 'required',
+            // 'description' => 'required',
+        ]);
+        
         $input = $request->except(['_token', '_method', 'document', 'college_id']);
 
         $mobility->update($input);
@@ -166,12 +204,16 @@ class MobilityController extends Controller
      */
     public function destroy(Mobility $mobility)
     {
+        $this->authorize('delete', Mobility::class);
+
         MobilityDocument::where('mobility_id', $mobility->id)->delete();
         $mobility->delete();
         return redirect()->route('mobility.index')->with('mobility_success', 'Your accomplishment in Inter-Country Mobility has been deleted.');
     }
 
     public function removeDoc($filename){
+        $this->authorize('delete', Mobility::class);
+
         MobilityDocument::where('filename', $filename)->delete();
         return true;
     }
